requirements.txt
opentelemetry-api~=1.20.0
opentelemetry-sdk~=1.20.0
opentelemetry-exporter-jaeger~=1.20.0
fastapi~=0.103.2